Remove only one form representation in edit mode
In edit mode, when editing lexemes with forms that have more than one
form representation, and clearing the text input for some of those
forms, only remove the form representation for the current language
code, not the whole form. Only remove the whole form if we would be
removing its last representation. (This holds regardless of whether the
currently edited language code is the “main” language code of the
template or a different one.)