Stop using mwoauth.identify() to get user info
This is a weird interface with an undocumented response format (I’m
pretty sure at the time I looked at the MediaWiki OAuth source code to
figure out what it returns), and it’s completely redundant with the
standard userinfo API, so let’s use that instead. This also applies the
T272319 “Nonce already used” workaround to these calls.