Use hyphens in i18n/ message keys
We continue to use underscores in Python, as it’s more pythonic, but use
hyphens in the JSON files, which is more in line with translatewiki.net
convention. The “key” variable(s) in the loading code intranslations.py
are renamed to clarify what kind of key they’re referring to.

JSON part done with the following command:

    for file in i18n/*.json; do
        jq --tab 'with_entries(.key |= gsub("_"; "-"))' "$file" \
	    | sponge "$file";
    done