Log API errors as if uncaught
The previous commit caused API errors to no longer be printed to the
debug log, so if the info in the HTML output wasn’t enough to deduce the
problem (or if the user never bothered to report about it), it would be
harder for me to fix it. Log the exception again by calling the right
method on the Flask app.