Test that translations only use allowed HTML elements
Since we emit the translations as HTML without further parsing, we
mustn’t allow them to contain HTML other than what we want (an <abbr>
for CSRF). I don’t want to have to remember that when reviewing each
translatewiki.net pull request, so let’s test it automatically.

The HTML tag syntax is based on the WhatWG HTML standard. [1]

[1]: https://html.spec.whatwg.org/multipage/syntax.html#elements-2