Allow incomplete gender format specs with fallback
Previously, the format spec for the !g conversion had to include values
for all three possible grammatical genders, otherwise the formatter
would crash when encountering an unlisted value. Change it to fall back
to "m" in that case. This can be useful when the English source message
does not require a gender conversion, but should still include one in
order to signal its availability to translators – in that case, the
fallback to "m" lets us avoid specifying the same string three times in
the format spec.