Add support for optional forms
Sometimes, certain forms of a template should not be required. This
could be done before by splitting off a separate template, or by
recommending advanced mode, but in some cases, it seems worthwhile to
support optional forms as first-class citizens. The implementation is
straightforward – just another condition in form2input() for omitting
the required attribute.

The initial template with optional forms is Norwegian Bokmål adjectives,
where the comparative and superlative forms are optional.