Improve language code handling
The externally visible impact of this change should be that we now tag
interface messages with the right language code if they’re not in the
overall interface language (because language fallback occured).
Internally, we now keep track of language codes a bit better: the new
language module contains methods to convert from lexicographical to
interface MediaWiki codes, and from those to either of HTML or Babel
codes, and we use those methods throughout the codebase. All entry
points (that potentially emit HTML; skipped for a few JSON-only APIs)
assign the flask.g.interface_language_code now, and we generally use
this to language-tag the HTML, instead of the template['language_code'].