Make formatters MarkupSafe-aware
Tweak the formatters so that they support interoperation with
MarkupSafe, assuming the markupsafe module is available, and correctly
escape values if the format string is Markup (but not if it is a plain
string). For CommaSeparatedListFormatter, this requires us to reach into
Babel internals, but let’s hope they stay stable for the time being.

Tests are included for various combinations of Markup and non-Markup
format strings and values.