Make /login redirect to referrer if opened directly
Once we tell users opening a template to log in via a link, rather than
redirecting them directly to the login page, the login page will still
need to know where to redirect them back afterwards. We can use the
referrer (HTTP Referer header) for this purpose.

If some other page (outside of this tool) hotlinks to our /login page,
that means we’ll eventually redirect the user back to that page. We
could try to avoid this by checking if the referrer URL starts with the
root URL for this tool, but that kind of comparison has been a constant
source of pain in the past (finally removed in f4e7ba98a7), so as long
as it’s not absolutely necessary, I’d rather avoid it. As far as I can
tell, redirecting to another page after login isn’t a security issue.

Part of #36.