Don’t redirect to login in edit mode
When loading edit mode, if the user isn’t logged in, instead of
redirecting to the login page, just show them a readonly version of the
page, with the inputs and button disabled and a message warning them
that they need to log in. This makes “edit mode” more useful as a lexeme
(forms) “viewer”.

That said, default contrast for disabled form inputs is apparently
pretty low, so I’m still thinking of adding a separate “view mode” that
would format the sentences differently (without inputs).

Part of #36.