Don’t redirect to login in bulk mode
This is mainly for consistency with the other two modes now – bulk mode
isn’t really reachable when not logged in anyways (at least not via any
links in the tool – of course users could have it bookmarked and then
have their session expire).

It does open up the possibility that the user will first see the
disabled form with the reminder to log in, only to log in as a
non-autoconfirmed account and be shown the rather ruder “not allowed”
message without any form at all. But I think that’s acceptable for now.
(Eventually, that message should probably be shown with the disabled
form as well.)

Part of #36.