Move login hint to top of page
On long templates, a lot of scrolling may be required before the bottom
of the form, where we previously put the login hint, becomes visible,
which means it’s not very obvious why the whole form is disabled. Moving
the main hint to the top of the page (where it’s now a <div> rather than
a <span>) solves that problem; to still give users an opportunity to
know why the button is disabled when they’re looking at it (and the hint
is potentially out of view), add a stripped-tags version of the message
as title text to the disabled button.

Follows-up #36.