Fix JS duplicates check for advanced mode
Previously, the JavaScript version of the check for potential duplicates
only checked the first input, whereas the server-side version accounts
for potentially empty forms in advanced mode, and uses the first
nonempty form representation variant as the lemma. I was aware of this
discrepancy at the time (see the last paragraph of the commit message of
f97fc25c67, the initial commit for findDuplicates.js), but didn’t fully
realize its impact: since the script starts out by removing the
server-generated duplicates warning (if it exists), this meant that it
was essentially impossible to create “duplicate” lexemes with the first
form missing without disabling JavaScript.

Fix this by actually watching for changes on all inputs, and using the
same logic to get the lemma as in the Python code. To prevent repeating
the search for the same lemma as changes are made in other forms, we
keep track of the last lemma, and only start a new search if the new
lemma is different; this also lets us check if the lemma is still the
same by the time the network request completes, hopefully avoiding
issues with interleaved network requests.

Fixes #44.