Cache noDuplicateHtml in JS
The JS code makes two network requests when checking for duplicates: one
to search for the duplicate, and one for the HTML of the “this is not a
duplicate” checkbox that it needs to add if a potential duplicate was
found. Neither of these requests was cached in any way, and for the
actual search, I think that’s okay – the search results aren’t likely to
change, but if they do, I think it’s nice to reflect that, and the same
search isn’t very likely to happen multiple times anyways.

But re-requesting the HTML for the checkbox every time is completely
pointless. Wrap that in a function and cache it, to cut down on requests
(and make the duplicates warning show up earlier!).

We have a similar safeguard as with the lastCheckedLemma – if multiple
requests for the noDuplicateHtml are kicked off, they all check upon
completion if the cachedNoDuplicateHtml is still null, so only the first
one will actually assign it, and the others will then return that (and
discard the result of their network request). There’s so strong need for
this – we could just as well

    return cachedNoDuplicateHtml = noDuplicateHtml;

– but it makes some sense to me to keep the HTML consistent, just in
case the server response varies for whatever reason.