Add lexeme ID to bulk mode result
Make submit_lexeme() return the lexeme ID in addition to the URI, and
show the ID in the bulk mode results.

(I decided against making submit_lexeme() return only the ID, and having
a separate function to turn it into a URL/URI, because such a function
would also have to take the wiki or host as input. In submit_lexeme(),
we already have that, so it makes sense to also build the URI there.)