Add template label to tool title
Introduce four messages for create, advanced, bulk, and edit mode, which
are used to include the template label in the tool’s <title>. We also
language-tag the <title> according to the template – the “Wikidata
Lexeme Forms” part is always in English, but it’s not possible to tag
only part of the title (it can’t contain any tags or child elements),
and the majority of the title (including the more important part) should
now be in the template language.

Part of #50.