Add lemmas and lexeme ID to title in edit mode
This effectively hard-codes the English word-separator and parentheses
messages, but so does the existing <h2> (and some other places), so I’ll
fix that at some other time. (As far as I’m aware, it’s not yet relevant
with any of the languages supported by the tool.)

Part of #50.