Add link to create duplicate lexeme in bulk mode results
When bulk mode refuses to create a lexeme because a potential duplicate
already exists, add a link to the regular page to create the lexeme
anyway. (On that page, users will then still have to check the “this is
not a duplicate” box – I don’t want to add a simple URL parameter to
pre-check that box because it could be abused, and guarding it with some
kind of token seems like too much effort.)

The previous “actionable” flag for the duplicates warning is replaced
with “in_bulk_mode”, since that’s what it means and determines how it’s
used. If we need more flexibility here later, then we’ll have to rename
the flag again and/or split it up, but for now this seems like the
simplest solution. (If there are two flags, and the “instructions” and
“create” messages can theoretically appear at the same time, then they
have to be separated somehow, but wrapping them each in a <p> adds
vertical whitespace…)

Fixes #51.