Make the navbar mobile-friendly
On narrow screens, the navbar wraps onto multiple lines. This causes it to be
collapsed by default instead. The HTML follows the usual Bootstrap style [1]
but since we're not loading Bootstrap's Javascript, this instead implements its
own simple click handler.

Bootstrap does not degrade nicely when Javascript is disabled, so a few
additional things are included so that it does:

* The classes to collapse the navbar are added by Javascript, so that it
  doesn't get collapsed with no way to expand it when Javascript isn't enabled.
* The toggle button is force-hidden via the `hidden` attribute by default which
  is then removed by Javascript, so that the toggle button is not shown when
  Javascript isn't enabled.
* Bootstrap doesn't display the navbar properly when there is a `<div>` with no
  collapse classes, so the initial no-JS-friendly state uses a `nojs` class
  which makes the navbar display properly before the Bootstrap classes are
  added.

[1]: https://getbootstrap.com/docs/4.6/components/navbar/