Add type declaration for Flask 2.1.0
Since Flask 2.1.0, mypy has started complaining about the

    app.session_interface.serializer.register(…)

code in app.py, because the declared type of session_interface is
SessionInterface, which doesn’t have a serializer member (that’s only
present on the subclass SecureCookieSessionInterface, which is the
default implementation). I’m not sure why this worked before (I didn’t
find a relevant commit in Flask), but let’s add the type declaration.