Update --extra-files option in README.md
Watching a whole directory for changes seems to have worked at the time
of commit 654f25e4cc, but it doesn’t work now and the Flask developers
don’t seem to be interested in supporting it either. Let’s just watch
the en.json file instead, that should be good enough for most cases.