Add venv/ to .gitignore and flake8 exclude
Seems like a no-brainer after we started using pip-tools (though I’m not
yet using the venv for most development on the tool… we’ll see about
that).

mypy doesn’t seem to check any venv files, so we don’t need to add
exclude = venv to the mypy.ini.