Add support for ambiguous template names
These are conceptually similar to redirects (i.e. string values in the
main templates dict), but there’s no single replacement, but rather a
list of them, which should be presented to the user. The initial
application will be Portuguese templates, but it may turn out to be
useful elsewhere as well.

test_app.py gains a test case for the existing feature of redirecting to
a renamed template; tests for ambiguous templates will be added once we
actually have any.