Test misleading capitalization in examples
If the example sentence has the placeholder at the beginning, then the
placeholder may have an initial uppercase letter even though the form
should not be entered this way; this can confuse users. (I believe this
was first observed for Portuguese templates.) I’ve sometimes tried to
watch out for that, but I’ve also failed to notice it in the past (see
commit 860abb205b fixing an old instance of this issue), so let’s add a
test for it.

The test is, for now, skipped for Armenian templates, since there are
some forms with this pattern there; I’ve pinged the template editors
on-wiki, and hopefully the template will soon be updated and then the
test can be unskipped for those templates.