Use flask.typing.ResponseReturnValue instead of ResponseValue
ResponseReturnValue includes tuples of response values with headers or
status codes, which we sometimes use. Import it under an abbreviated
name for readability (I think an acronym is more readable than a very
long name in this case).

Along with this, add typing to match_template_to_lexeme_id(), the first
function that needs ResponseReturnValue instead of ResponseValue as the
return type; to make that work, also ensure that it only gets real
templates when resolving redirects (which is checked by a unit test).