Recompile requirements.txt on Python 3.9
Apparently this depends on the Python version. I use 3.10 at home, but
in Toolforge we only have 3.9, and apparently under 3.10 pip-tools
throws away some dependencies that are still needed under 3.9. (Or this
might be due to distribution rather than version, i.e. the fact that
it’s Debian on Toolforge… I’m not sure, but after all, the 3.10
requirements.txt worked in CI, which is also 3.9.)