Don’t call get_gender() when not needed
If {{GENDER:}} is used in a message, but only one replacement is
specified (e.g. {{GENDER:$1|You}} rather than {{GENDER:$1|he|she}} or
{{GENDER:$1|he|she|they}}), then it’s pointless to call get_gender(),
because we’ll just fall back to the first and only replacement anyways.
Short-circuit that case to avoid the (potential) API call.

This is especially relevant for English messages, which can contain
unused {{GENDER:}} as a hint to translators that the feature is
available; in other languages, I generally remove unused {{GENDER:}}
again when translators add it.