Move dev requirements to separate files
Following the instructions at [1]. This cuts the size of the production
venv in half, with the lion’s share of the savings coming from mypy.

The pip-sync way to install dependencies is now:

    pip-sync *requirements.txt

At some point I should probably make the README and/or GitHub actions
use pip-tools instead of bare pip.

[1]: https://pip-tools.readthedocs.io/en/latest/#workflow-for-layered-requirements