Handle missing lemmas in bulk-result.html
If we’re editing an existing lexeme, we currently don’t have any lemmas
to display, because build_lexeme() doesn’t build lemmas when a lexeme ID
is given (which also means we never update the lemma – note that this is
unlike edit mode). I don’t think I want to change build_lexeme() at the
moment (though note that it does already get the existing lexeme data,
for matching, so we theoretically have the original lemmas available),
so just update the template to cope with missing lexemes.