Add Hindustani non-verb templates
These templates were provided by [[User:عُثمان]] [1] at
[[Wikidata:Wikidata Lexeme Forms/Hindustani]] [2] – thank you!
(Specifically, as of [[Special:PermanentLink/1842047491]] [3].)

These templates come in two versions, Hindi (hi) and Urdu (ur). The two
are standardized registers of the broader Hindustani language, and
apparently close enough that it makes sense to reuse the templates, but
the writing systems are different (even in directionality); for a while
we entertained the thought of supporting multiple scripts within one
template, but this seems easier overall. A new test is added to verify
that the templates are identical apart from the labels and examples.

The verb templates are still missing. There’s nothing wrong with them
(as far as I know) – I’ll just need more time to transcribe them, which
I don’t have tonight.

[1]: https://www.wikidata.org/wiki/User:%D8%B9%D9%8F%D8%AB%D9%85%D8%A7%D9%86
[2]: https://www.wikidata.org/wiki/Wikidata:Wikidata_Lexeme_Forms/Hindustani
[3]: https://www.wikidata.org/wiki/Special:PermanentLink/1842047491