Support somevalue/novalue snaks
This will be needed for the upcoming Hindustani verb templates, where
some forms have grammatical aspect no value. mypy alerted me that
matching.py needed an update, which was nice; also add tests to
cover the new logic.

statement() / statements() in templates.py only supports novalue, not
somevalue, because we only need no value for now. If we ever need
unknown value as well, I’ll reconsider if it needs a shortcut (perhaps
with special marker values instead of None). Add somevalue support to
the types, though, because it’s easy to do.