Fix hindustani-verb-additive-causative-double-ur label
I must’ve accidentally not copied the label, so when I went to paste the
label, I accidentally pasted the identifier again instead. Add a test to
guard against this case as well. (It’s also happened occasionally with
the form labels and examples in the past, but we have tests to catch
that, examples must have a valid placeholder and labels must not be
valid examples.)