Add entity_ids package with property_ids module
This will be used to replace the hard-coded entity IDs in templates.py
(and matching.py) with variables. That should make the templates not
only much more readable in the source code, but hopefully also easier to
transcribe, as typing (prefixes of) variables named after labels should
be easier than typing (prefixes of) entity IDs.

The property_ids module is also executable, and prints CSS to stdout if
run as the main module: this can be added to the user styles, and will
replace entity IDs on [[Wikidata:Wikidata Lexeme Forms]] [1] subpages
with the variable names, to assure the transcriber that the referenced
item ID really corresponds to a named variable and doesn’t just happen
to have a similar label. The whole entity_ids package is executable as
well, running all the packages inside (currently just property_ids, but
more will follow). Note that we have to run each package as a separate
process, so that they all have their isolated globals().

dump_entity_ids.py is based on test_entities_exist() in
test_templates.py, and used to initially populate property_ids.py; it
will later be adjusted for other types of entity IDs, and ultimately
removed. (After the initial import of entity IDs, further entity IDs are
expected to be added manually as needed, but I’m not going to add the
200+ already existing entity IDs by hand.)

[1]: https://www.wikidata.org/wiki/Wikidata:Wikidata_Lexeme_Forms