Fix CI
I forgot to run it before pushing the previous commit, and now I don’t
want to amend it because I already mentioned the commit ID in an edit
summary on Wikidata. The fixes for the type errors are straightforward,
if a bit annoying; I have no idea how the whitespace error snuck in.