Test Python 3.11 in CI
We’re probably going to get Python 3.11 in Toolforge soon after 10 June
(T335507), so let’s start testing it. Note that Arch Linux is still on
Python 3.10 (though I heard an update is in the works), so this version
is *only* tested in CI for now, and also we can’t use Python 3.11-only
features quite yet.