Add test for optional forms coming last in a section
This condition is true for most templates, and when it’s false it’s
often because I forgot to add 'optional': True when transcribing a
template. (This happened several times in the Hindustani and Punjabi
templates, but was so far thankfully caught by the tests checking that
the templates match betwen the two language codes, since I wouldn’t
usually make the same random mistake twice. I think those templates are
also more prone to the mistake since they put “optional” in the form
“heading” after the form’s number at the beginning of its section, where
I’m liable to forget it by the time I reach the end of the form in the
Python code where I put 'optional': True after everything else; in some
other templates, the “optional” indication is instead at the end of a
form’s section, as in the Python code.)

There are a few exceptions, template names where we skip the test. In
Malayalam nouns, and possibly in Norwegian Bokmål masculine/neuter nouns
too, you could arguably add a section break between the optional and
next non-optional form (for Malayalam it’s between singular and plural
forms, eight each; for Bokmål the separation is a bit less clear to me),
but for now let’s just leave the templates alone, I think.