Make most of the tool translatable
This changes the way the HTML is language-tagged. Previously, the <html>
was hard-coded to lang="en", matching the hard-coded English text in the
navbar, while the <main> had the interface language code, which was also
the same as the template language code. Now, the <html> has the
interface language code, which is no longer the template language code
(unless they happen to be the same), and the template contents are
language-tagged only at the <dl> level. (The <main> contains other
content which is in the interface language: duplicate warnings and the
like.) To match the language code declared on the <html> and make the
whole tool translatable, new messages for the navbar elements are
introduced. (There are still a couple of hard-coded English texts that I
might fix later, see `git grep [A-Z] 'templates/*.html'`.)

The “optional” span after optional forms is currently not
language-tagged correctly because add_lang_if_needed() only checks based
on the user interface language code and doesn’t know that there’s a
different lang= (from the template) surrounding this particular message.
This will be fixed in a separate commit.