Use - instead of _ in message keys
It made sense to use _ in message keys when the translations were still
hand-written Python dicts, as _ is more Pythonic; and it made sense to
keep the message keys this way, and change the MediaWiki-like - to _
on-the-fly during the transition to translatewiki.net; but it doesn’t
make sense to keep the mixture of _ and - now, well after the transition
is over – it just adds confusion when the “same” message keys are
written two different ways. Move everything to the - form now.

I tried to find all the places that needed updating with:

    git grep "message\(_with_kwargs\)\?\s*(\s*[\"'][^\"']*_"