Unskip other translations
Skipping these made sense back when the user interface language was tied
to the template language, so the translations wouldn’t have been
reachable anyways. But now that the user interface language setting
exists, the translations are in principle usable, and it seems a shame
to throw them away; instead, let’s try a bit harder to make the language
codes palatable to Babel. For most of these I just followed the
MediaWiki fallbacks (as specified in languages/messages/Messages*.php)
without investigating in detail which e.g. plural forms both languages
have in MediaWiki.