Add Babel language code of skr-arab (Saraiki)
Strictly speaking, this would change a hypothetical skr-latn or other
scripts too, because we only look up the part of the code before the
hyphen in the dict. I should probably change that soon. (But there’s
currently no other skr-* in MediaWiki as far as I can tell, so it’s not
terribly urgent I think.)