Remove unneeded {{GENDER:}} from logged-in translations
These broke the tool: when the Python message looks like

    {user_name!g:m=Logged in as {user_link}}

then the user_link is interpolated before the user_name gender spec is
interpreted, which means the spec sees a https://a s part of the m=...
content and thinks the : is another separator there, truncating the
user_link there (in the middle of an attribute and tag). I don’t really
see how this can be fixed in the formatter, so let’s just update the
translations and add a test to ensure it doesn’t happen with any future
translations, I think.