Add separators in Dutch noun templates
The Dutch noun templates currently display the main forms in the left column
and diminutive forms in the right column. Adding separators will cause the
singular forms to be in the left column and plural forms to be in the right
column, matching the Dutch Wiktionary [1]. This is also more consistent with
the German templates (as a neighbouring and closely related language, it would
not be surprising if a Dutch speaker also wants to edit German). The tab order
will not be affected.

[1] e.g. https://nl.wiktionary.org/wiki/appelsien