Remove i18n/zh.json
Replaced by (currently identical) i18n/zh-hans.json. See also:
https://gerrit.wikimedia.org/r/c/translatewiki/+/935503

Add a special case to the test to allow 'zh' translations to be missing
even though we have 'zh' templates. (The test is no longer as important
as it used to be, since the template language is no longer used for the
user interface language; that said, I think it’s still useful to keep in
general.)