Add first Wikifunctions for German neuter nouns
-en/-um isn’t the most common case, but it’s relatively simple. More
will follow as part of #140.

However, this does require support for slashes in the function name, so
swap the function name and lemma around in the API URL and instruct
Flask to parse the function name as a “path” (meaning it can contain
slashes) rather than the lemma. Note that the frontend JS code was
already splitting the input on slashes and only sending the first one to
the API, so there was no point for the backend to support slashes there;
if there really is a use case for slashes in the very first form (I
doubt it, to be honest – ffb3d07f3e didn’t really justify it), then I
suppose we can handle it in the JS code by just making several
consecutive API calls.