Rename message variables to consistently match type
The main goal of this is to be able to test these messages more
systematically (see 90166527d9), for which the tests need to know the
type of each parameter (how we expect it to be used in each
translation); originally I thought about encoding this using wrapping
dataclasses or something similar:

    # ...
    'description-with-forms-and-senses': [StrArg('description'), PluralArg('forms'), PluralArg('senses')],
    'bulk-not-allowed': [GenderArg('user')],
    # ...

But while converting the specification to this syntax, I realized that
the information was often already contained in the variable name, and
the cases where it wasn’t could be renamed accordingly. This also makes
the message calls in *.py and *.html files much more readable.

We don’t do anything with this information just yet, that’ll come in
separate commits.