Let Wikifunctions infer first form from lemma in edit mode
This tool assumes that the first nonempty form is identical to the
lemma. (There is a stale branch, `lemma`, which lets templates select
non-first form for the lemma; it’s awaiting review / verification.) If a
lexeme is missing all forms in edit mode, it’s reasonable to assume that
the first form can be initialized from the lemma; so far we didn’t do
this automatically, so that edit mode, without any user action, doubles
as a *viewer* of the existing data matched to the template. But with
Wikifunctions support, we can infer the first form from the lemma once
the user has actively clicked one of the Wikifunctions buttons; the
other forms can then be generated from the first one as usual.

Fixes #199.