Configure Toolforge health-check-path
Previously, we manually patched the deployment created by `webservice`
to add a startup probe. Now we can instead declare the health-check-path
in service.template and Toolforge will automatically create a startup
and liveness probe for us. This also means that we can finally restart
the tool with just `webservice restart` again (previously, this either
didn’t perform a rolling restart at all or did one but wiped out the
startup probe in the process).

Because we now have a liveness probe in addition to a startup probe, the
/healthz endpoint gets a lot more requests than before; fortunately it
turns out to be pretty easy to exclude it from the uWSGI logs [1][2].

[1]: https://stackoverflow.com/a/25686557/1420237
[2]: https://uwsgi-docs.readthedocs.io/en/latest/InternalRouting.html

Bug: T341919