Test message HTML more comprehensively
Previously, we didn’t test attributes at all, so a malicious translation
could have had something like

    <abbr onmouseenter="alert('pwnd')">Create</abbr>

and executed arbitrary JavaScript. (Fortunately, no translation actually
did this, as is evidenced by the new tests still passing.)

In theory I probably could’ve still tested this with an HTML-based regex
parser – the basic element + attribute syntax isn’t *that* complicated –
but it seems nicer to use a library dedicated to it.

(allowed_global_attributes is needed because the fa translation of the
CSRF message sets lang= and dir= on the CSRF <abbr>. Nice ^^)