Load translations separately in tests
When these tests are moved to a separate package / module, they won’t
have access to the app anymore.

I originally intended to make the TranslationsConfig a test fixture that
the tool would have to provide (as mentioned in e530449d5e), but I don’t
think that works out – it’s impossible to parametrize tests based on
fixtures (https://stackoverflow.com/a/56865893/1420237). Instead, let’s
just demand that tools put the config in a well-known package name that
we can import (tool_translations_config).