Move most of language_info to toolforge_i18n
fallbacks() is needed in message() (which should also move to
toolforge_i18n); autonym() isn’t directly needed, but uses the same
underlying API request, and it seems likely enough that other tools will
also have a use for it (language selector), so I think it’s reasonable
to include it. label(), on the other hand, is quite specific to this
tool.

test_language_info will have to change a bit once this actually becomes
a separate package (at which point, if I understand correctly, the
relative import will stop working, but also we won’t have the conflict
with the top-level language_info anymore; also I should put the git repo
in the user agent then), but for now it works. Note that, to make the
tests work when they’re run on their own, we have to set an explicit
user_agent. (If we run all the tests at once, skipping this step still
works because then something else will pull in app with its
toolforge.set_user_agent() call.)