Replace internal WikiLambda API with public one
It works mostly the same, but the outer 'query' layer is gone (yay), the
parameter name is no longer prefixed (yay), and the action begins with
“wikifunctions” instead of “wikilambda” (eh? but I don’t see this tool
using another WikiLambda install any time soon, so, fine). The 'success'
member is also gone, but I think that was mostly pointless anyway
(surely the API should respond with a proper error, which mwapi can
detect and throw automatically, if anything is wrong), so we can just
check the 'data' instead.

Bug: T360374