Remove no longer needed legacy_language_codes entry
It’s been over five months, I think that’s long enough for any users
with 'yue' to have visited the tool and have their preferences
automatically updated to 'zh-hant'. Leave an explanatory comment in the
dict instead.

Add a type to the dict so mypy is happy (can’t infer a type from an
empty dict).