Remove skipped_language_codes, load documentation
The skipped_language_codes general mechanism was introduced in
6231e4139a and became mostly unused again in db3dd67b8a (see those
commits for rationale); I don’t think we need it as a general mechanism
anymore. Instead, make load_translations load the message documentation
separately (without converting it into format strings), and return it so
we can use it for tests later.