Fix unused imports and Flake8 config
Flatly disabling the Flake8 warnings about unused imports is evidently
not a good idea – it’s just not tenable for me to stay on top of the
imports manually while refactoring. Instead, we can just disable those
Flake8 rules on the relevant import lines. (F405 remains disabled
because it otherwise produces tens of thousands of violations that we
would have to disable on virtually every line of templates.py.)