Add support for ?uselang=qqx
Using a pretty similar format to MediaWiki, though arguments are shown
as their repr() – this means strings are wrapped in single quotes (which
is a bit unnecessary), but also means Markup() is visible, and remember
that message arguments can also be lists.

I don’t think we should treat qqx as a real language code, though, so
?uselang=qqx actually sets the interface language to English and then
sets a separate flag in flask.g that the message() function reads. So
the HTML is still in lang="en" throughout.